+++ includes/class-tmsm-frontend-optimizations.php
@@ -253,8 +253,8 @@ class Tmsm_Frontend_Optimizations {
 		// WP Rocket
 		$this->loader->add_filter( 'rocket_lazyload_excluded_attributes', $plugin_public, 'rocket_lazyload_excluded_attributes_elementor', 10 );
 
-		// GDPR Data Request Recipient
-		$this->loader->add_filter( 'user_request_confirmed_email_to', $plugin_public, 'user_request_confirmed_email_to', 10 );
+		// Personal Data Request Recipient
+		$this->loader->add_filter( 'user_request_confirmed_email_to', $plugin_public, 'user_request_confirmed_email_to_dpo', 10, 2 );
 
 		// WooCommerce
 		$this->loader->add_action( 'woocommerce_scheduled_sales', $plugin_public, 'woocommerce_scheduled_sales_empty_cache', 200 );