+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Fix DPO recipient function fatal error
+
 ### 1.0.10: April 1st, 2019
 * Remove Jetpack Geo Location module